ScrollView added to main activity

+++ UVMe - No Fragments/UVMe/app/src/main/java/com/elec_coen_390/uvme/BatteryData.java
@@ -2,17 +2,17 @@ package com.elec_coen_390.uvme;
 
 public class BatteryData {
 
-    private static float batteryLevel = 0;
+    private static int batteryLevel = 0;
 
-    public BatteryData(float battery_level) {
+    public BatteryData(int battery_level) {
         batteryLevel = battery_level;
     }
 
-    public static float getBatteryLevel() {
+    public static int getBatteryLevel() {
         return batteryLevel;
     }
 
-    public static void setBatteryLevel(float battery_level) {
+    public static void setBatteryLevel(int battery_level) {
         batteryLevel = battery_level;
     }
 }
